feat: 레이스실행main 구현
